Add xz version scan script and update .gitignore
A bash script, xz-scan.sh, has been added that checks if the system is running vulnerable versions of the xz package and conducts a scan of all executables. Also, the .gitignore file has been updated to include ".idea" folder. This script uses package managers dpkg, rpm, and pacman to pull the xz version and provides output whether the version is vulnerable, patched, or undetermined.